+++ tests/integration/annotation_import/test_mea_prediction_import.py
@@ -58,6 +58,7 @@ def test_create_from_objects_with_confidence(predictions_with_confidence,
     annotation_import_test_helpers.assert_file_content(
         annotation_import.input_file_url, predictions_with_confidence)
     annotation_import.wait_until_done()
+    assert annotation_import.state == AnnotationImportState.FINISHED
     annotation_import_test_helpers.download_and_assert_status(
         annotation_import.status_file_url)
 
@@ -80,6 +81,7 @@ def test_create_from_objects_all_project_labels(
 def test_model_run_project_labels(model_run_with_all_project_labels,
                                   model_run_predictions):
     model_run = model_run_with_all_project_labels
+    # TODO: Move to export_v2
     model_run_exported_labels = model_run.export_labels(download=True)
     labels_indexed_by_schema_id = {}
     for label in model_run_exported_labels: