+++ tests/integration/test_batch.py
@@ -1,10 +1,16 @@
 import time
-
-from labelbox.exceptions import ProcessingWaitTimeout
 import pytest
+
+from uuid import uuid4
 from labelbox import Dataset, Project
+from labelbox.exceptions import ProcessingWaitTimeout, MalformedQueryException, ResourceConflict, LabelboxError
 
 IMAGE_URL = "https://storage.googleapis.com/diagnostics-demo-data/coco/COCO_train2014_000000000034.jpg"
+EXTERNAL_ID = "my-image"
+
+
+def get_data_row_ids(ds: Dataset):
+    return [dr.uid for dr in list(ds.export_data_rows())]
 
 
 @pytest.fixture
@@ -12,20 +18,7 @@ def big_dataset(dataset: Dataset):
     task = dataset.create_data_rows([
         {
             "row_data": IMAGE_URL,
-            "external_id": "my-image"
-        },
-    ] * 2)
-    task.wait_till_done()
-
-    yield dataset
-
-
-@pytest.fixture
-def small_dataset(dataset: Dataset):
-    task = dataset.create_data_rows([
-        {
-            "row_data": IMAGE_URL,
-            "external_id": "my-image"
+            "external_id": EXTERNAL_ID
         },
     ] * 3)
     task.wait_till_done()
@@ -55,6 +48,81 @@ def test_create_batch(project: Project, big_dataset: Dataset):
     batch = project.create_batch("test-batch", data_rows, 3)
     assert batch.name == "test-batch"
     assert batch.size == len(data_rows)
+
+
+def test_create_batch_with_invalid_data_rows_ids(project: Project):
+    with pytest.raises(MalformedQueryException) as ex:
+        project.create_batch("test-batch", data_rows=['a', 'b', 'c'])
+        assert str(
+            ex) == "No valid data rows to be added from the list provided!"
+
+
+def test_create_batch_with_the_same_name(project: Project,
+                                         small_dataset: Dataset):
+    batch1 = project.create_batch("batch1",
+                                  data_rows=get_data_row_ids(small_dataset))
+    assert batch1.name == "batch1"
+
+    with pytest.raises(ResourceConflict):
+        project.create_batch("batch1",
+                             data_rows=get_data_row_ids(small_dataset))
+
+
+def test_create_batch_with_same_data_row_ids(project: Project,
+                                             small_dataset: Dataset):
+    batch1 = project.create_batch("batch1",
+                                  data_rows=get_data_row_ids(small_dataset))
+    assert batch1.name == "batch1"
+
+    with pytest.raises(MalformedQueryException) as ex:
+        project.create_batch("batch2",
+                             data_rows=get_data_row_ids(small_dataset))
+        assert str(ex) == "No valid data rows to add to project"
+
+
+def test_create_batch_with_non_existent_global_keys(project: Project):
+    with pytest.raises(MalformedQueryException) as ex:
+        project.create_batch("batch1", global_keys=["key1"])
+        assert str(
+            ex
+        ) == "Data rows with the following global keys do not exist: key1."
+
+
+def test_create_batch_with_negative_priority(project: Project,
+                                             small_dataset: Dataset):
+    with pytest.raises(LabelboxError):
+        project.create_batch("batch1",
+                             data_rows=get_data_row_ids(small_dataset),
+                             priority=-1)
+
+
+def test_create_batch_with_null_priority(project: Project,
+                                         small_dataset: Dataset):
+    with pytest.raises(LabelboxError):
+        project.create_batch("batch1",
+                             data_rows=get_data_row_ids(small_dataset),
+                             priority=0)
+
+    with pytest.raises(LabelboxError):
+        project.create_batch("batch1",
+                             data_rows=get_data_row_ids(small_dataset),
+                             priority=None)
+
+
+def test_create_batch_with_out_of_bound_priority(project: Project,
+                                                 small_dataset: Dataset):
+    with pytest.raises(LabelboxError):
+        project.create_batch("batch1",
+                             data_rows=get_data_row_ids(small_dataset),
+                             priority=6)
+
+
+def test_create_batch_with_float_number_priority(project: Project,
+                                                 small_dataset: Dataset):
+    with pytest.raises(LabelboxError):
+        project.create_batch("batch1",
+                             data_rows=get_data_row_ids(small_dataset),
+                             priority=4.9)
 
 
 def test_create_batch_async(project: Project, big_dataset: Dataset):
@@ -125,7 +193,7 @@ def test_batch_creation_for_data_rows_with_issues(
     ]
     data_rows_to_add = valid_data_rows + invalid_data_rows
 
-    assert len(data_rows_to_add) == 5
+    assert len(data_rows_to_add) == 4
     batch = project.create_batch("batch to test failed data rows",
                                  data_rows_to_add)
     failed_data_row_ids = [x for x in batch.failed_data_row_ids]
@@ -165,7 +233,7 @@ def test_export_data_rows(project: Project, dataset: Dataset):
     task = dataset.create_data_rows([
         {
             "row_data": IMAGE_URL,
-            "external_id": "my-image"
+            "external_id": EXTERNAL_ID
         },
     ] * n_data_rows)
     task.wait_till_done()
@@ -179,6 +247,51 @@ def test_export_data_rows(project: Project, dataset: Dataset):
 
     assert len(result) == n_data_rows
     assert set(data_rows) == set(exported_data_rows)
+
+
+def test_list_all_batches(project: Project, client):
+    """
+    Test to verify that we can retrieve all available batches in the project.
+    """
+    # Data to use
+    img_assets = [{
+        "row_data": IMAGE_URL,
+        "external_id": str(uuid4())
+    } for asset in range(0, 2)]
+    data = [img_assets for _ in range(0, 2)]
+
+    # Setup
+    batches = []
+    datasets = []
+
+    for assets in data:
+        dataset = client.create_dataset(name=str(uuid4()))
+        create_data_rows_task = dataset.create_data_rows(assets)
+        create_data_rows_task.wait_till_done()
+        datasets.append(dataset)
+
+    for dataset in datasets:
+        data_row_ids = get_data_row_ids(dataset)
+        new_batch = project.create_batch(name=str(uuid4()),
+                                         data_rows=data_row_ids)
+        batches.append(new_batch)
+
+    # Test
+    project_batches = list(project.batches())
+    assert len(batches) == len(project_batches)
+
+    for project_batch in project_batches:
+        for assets in data:
+            assert len(assets) == project_batch.size
+
+    # Clean up
+    for dataset in datasets:
+        dataset.delete()
+
+
+def test_list_project_batches_with_no_batches(project: Project):
+    batches = list(project.batches())
+    assert len(batches) == 0
 
 
 @pytest.mark.skip(